requirements.txt
amqp==2.1.4

Babel==2.3.4
billiard==3.5.0.2
celery==4.0.2
Django==1.10.5
django-celery-results==1.0.1
djangorestframework==3.5.4
flower==0.9.1
kombu==4.0.2
pytz==2016.10
tornado==4.2
vine==1.1.3
wheel==0.26.0
